requirements.txt
amqp>=5.2.0
asgiref>=3.8.0
async-timeout>=4.0.3
billiard>=4.2.0
celery>=5.3.6
click>=8.1.7
click-didyoumean>=0.3.0
click-plugins>=1.1.1
click-repl>=0.3.0
colorama>=0.4.6
Django==5.2.5
django-environ>=0.11.2
djangorestframework>=3.15.0
et-xmlfile>=1.1.0
greenlet>=3.0.3
gunicorn>=21.2.0
kombu>=5.3.5
numpy>=1.26.3
numpy-financial>=1.0.0
openpyxl>=3.1.2
pandas>=2.2.0
prompt-toolkit>=3.0.43
psycopg2-binary>=2.9.9
pyarrow>=15.0.0
python-dateutil>=2.8.2
python-decouple>=3.8
pytz>=2023.3
redis>=5.0.1
six>=1.16.0
SQLAlchemy>=2.0.25
sqlparse>=0.4.4
typing_extensions>=4.9.0
tzdata>=2023.4
vine>=5.1.0
wcwidth>=0.2.13